Bug fix: CZ phases
As written originally, the CZ phases for *all* two-qubit gates were simply updated to be those for the last defined gate. That obviously doesn't work! Now, when exploding composite gates, we first check to see if the gate being exploded is a CPHASE_with_1qb_phases - type gate, and if so updates it with the proper phases.

This is really not an ideal approach. More ideal would be to define not just a _pulse_ that's specific for each qubit, but also a set of _gates_ that are specific for each qubit. But perhaps that's too hefty and these couple lines of code are sufficient.